requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
distlib==0.3.5
Django==4.1
djangorestframework==3.13.1
filelock==3.8.0
idna==3.4
panda==0.3.1
Pillow==9.2.0
pipenv==2022.8.24
platformdirs==2.5.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.16.3
virtualenv-clone==0.5.7
